refactor(web): split the monolithic api/client.ts by domain
client.ts was a ~1,580-line god-module holding the shared fetch core, the
single `api` object (every endpoint method), and every request/response
type. Split it by domain into sibling modules under web/src/api/ behind a
re-export barrel so the ~70 existing importers keep working unchanged.

- core.ts: BINDERY_BASE, CSRF state + initCSRF, request()/uploadFile(),
  ApiError, isNoDownloadClientError (the shared private foundation).
- common.ts: shared Page/PaginatedResponse envelopes + BookRef.
- One module per domain (system, auth, books, authors, bulk, indexers,
  downloadclients, library, queue, history, notifications, profiles,
  series, settings, calibre, grimmory, abs, importlists, recommendations),
  each exporting its types and a partial `*Api` slice.
- client.ts is now a 91-line barrel: re-exports ApiError/BINDERY_BASE/
  isNoDownloadClientError/initCSRF from core, re-exports every domain's
  types via `export type *`, and composes `api` from the partial slices.

No behavior, request shapes, type definitions, or exported names changed.
The 99 original public symbols all resolve identically. CSRF mutable state
stays in one module (core) so request() and the auth flows share one source
of truth, preserving the original module-closure semantics.

Verified: typecheck clean, build succeeds, lint has no new warnings (none
in api/), all 266 tests pass.

Refs #547, #1000.
